+++ test/compute-engine/latex-syntax/matchfix.test.ts
@@ -114,14 +114,14 @@ describe('MATCHFIX abs and norm', () => {
     expect(check('|(1+|a|+2)|')).toMatchInlineSnapshot(`
       box       = ["Abs", ["Delimiter", ["Add", 1, ["Abs", "a"], 2]]]
       canonical = ["Abs", ["Add", ["Abs", "a"], 1, 2]]
-      simplify  = ||a| + 3|
+      simplify  = |a| + 3
     `));
 
   test('|1+|a|+2|', () =>
     expect(check('|1+|a|+2|')).toMatchInlineSnapshot(`
       box       = ["Abs", ["Add", 1, ["Abs", "a"], 2]]
       canonical = ["Abs", ["Add", ["Abs", "a"], 1, 2]]
-      simplify  = ||a| + 3|
+      simplify  = |a| + 3
     `));
 
   test('||a||', () =>