+++ test/compute-engine/smoke.test.ts
@@ -389,20 +389,9 @@ describe('CANONICALIZATION multiply', () => {
     expect(canonicalToJson('2\\times\\frac12')).toMatchInlineSnapshot(`1`));
 
   test(`2\\times(5-5)\\times5\\times4`, () =>
-    expect(canonicalToJson('2\\times(5-5)\\times5\\times4'))
-      .toMatchInlineSnapshot(`
-      [
-        Multiply,
-        2,
-        4,
-        5,
-        [
-          Add,
-          -5,
-          5,
-        ],
-      ]
-    `));
+    expect(
+      canonicalToJson('2\\times(5-5)\\times5\\times4')
+    ).toMatchInlineSnapshot(`0`));
 
   test(`(-2)\\times(-x)\\times y\\times\\frac{3}{-5}`, () =>
     expect(canonicalToJson('(-2)\\times(-x)\\times y\\times\\frac{3}{-5}'))